+++ R/weighting.R
@@ -1,27 +1,28 @@
 ##' Weighted mean and variance of a vector
 ##'
-##' Compute the weighted mean or weighted variance of a vector.
+##' Compute the weighted mean or weighted variance of a vector. Exact copies of Hmisc 
+##' functions.
 ##' 
 ##' @aliases wtd.var
 ##' @param x Numeric data vector 
 ##' @param weights Numeric weights vector. Must be the same length as \code{x}
-##' @param normwt Only for \code{wtd.var}, if \code{TRUE} then weights are normalized for the weighted count to be the same as the non-weighted one
+##' @param normwt Only for \code{wtd.var}, if \code{TRUE} then weights are normalized 
+##' for the weighted count to be the same as the non-weighted one
 ##' @param na.rm if \code{TRUE}, delete \code{NA} values.
 ##' @details
 ##' If \code{weights} is \code{NULL}, then an uniform weighting is applied.
 ##' @author
 ##' These functions are exact copies of the \code{wtd.mean} and \code{wtd.var}
-##' function from the \link[Hmisc]{wtd.stats} package. They have been created by Frank Harrell, Department of Biostatistics,
-##' Vanderbilt University School of Medicine, <f.harrell@@vanderbilt.edu>.
+##' function from the \link[Hmisc]{wtd.stats} package. They have been created by 
+##' Frank Harrell, Department of Biostatistics, Vanderbilt University School of 
+##' Medicine, <f.harrell@@vanderbilt.edu>.
 ##' @seealso
 ##' \code{\link{mean}},\code{\link{var}}, \code{\link{wtd.table}} and the \code{survey} package.
 ##' @examples
 ##' data(hdv2003)
 ##' mean(hdv2003$age)
 ##' wtd.mean(hdv2003$age, weights=hdv2003$poids)
-##' var(hdv2003$age)
-##' wtd.var(hdv2003$age, weights=hdv2003$poids)
-##' @export wtd.mean wtd.var
+##' @export wtd.mean
 
 `wtd.mean` <-
 function (x, weights = NULL, normwt = "ignored", na.rm = TRUE) 
@@ -36,24 +37,6 @@ function (x, weights = NULL, normwt = "ignored", na.rm = TRUE)
     sum(weights * x)/sum(weights)
 }
 
-`wtd.var` <-
-function (x, weights = NULL, normwt = FALSE, na.rm = TRUE) 
-{
-    if (!length(weights)) {
-        if (na.rm) 
-            x <- x[!is.na(x)]
-        return(stats::var(x))
-    }
-    if (na.rm) {
-        s <- !is.na(x + weights)
-        x <- x[s]
-        weights <- weights[s]
-    }
-    if (normwt) 
-        weights <- weights * length(x)/sum(weights)
-    xbar <- sum(weights * x)/sum(weights)
-    sum(weights * ((x - xbar)^2))/(sum(weights) - 1)
-}
 
 
 #' Weighted one-way and two-way frequency tables.
@@ -82,7 +65,7 @@ function (x, weights = NULL, normwt = FALSE, na.rm = TRUE)
 #' wtd.table(hdv2003$sexe, weights=hdv2003$poids, normwt=TRUE)
 #' table(hdv2003$sexe, hdv2003$hard.rock)
 #' wtd.table(hdv2003$sexe, hdv2003$hard.rock, weights=hdv2003$poids)
-#' @export
+#' @export 
 
 
 
@@ -123,6 +106,11 @@ function (x, y = NULL, weights = NULL, digits = 3, normwt = FALSE, na.rm = TRUE,
   as.table(result)
 }
 
+##' @export
+
+wtd.var <- function() {
+  stop("questionr::wtd.var has been removed. Please use Hmisc::wtd.var instead.")
+}
 
 #' Weighted Crossresult
 #'